+++ src/DotNetWorker.Core/Hosting/WorkerOptions.cs
@@ -21,5 +21,11 @@ namespace Microsoft.Azure.Functions.Worker
         /// Gets the collection of input converters.
         /// </summary>
         public InputConverterCollection InputConverters { get; } = new InputConverterCollection();
+
+        /// <summary>
+        /// Gets and sets the flag for opting in to unwrapping user-code-thrown
+        /// exceptions when they are surfaced to the Host. 
+        /// </summary>
+        public bool EnableUserCodeException { get; set; } = false;
     }
 }