+++ release_notes.md
@@ -3,5 +3,6 @@
 - My change description (#PR/#issue)
 -->
 
+- Added support for surfacing user-thrown exception to App Insights (#939)
 - Add new analyzer rules for the source-generated function metadata preview (#1048)
 - Bump gRPC dependencies to latest version (#1085)